generate_pssh2.pl can now use md5sum and sequence file (if present), it also check the pp cache before running hhblits

+++ src/pssh2/generate_pssh2.pl
@@ -9,6 +9,7 @@
 use strict;
 use File::chdir;
 use Getopt::Long;
+use POSIX;
 
 # PARAMETERS
 my $rootDir = "/mnt/project/pssh/pssh2_project/";
@@ -17,17 +18,22 @@ my $queries_dir = $rootDir."data/uniprot_derived/sprot_fastas";
 my $localHhblitsDir = "/var/tmp/rost_db/data/hhblits/";
 my $uniprot20 = $localHhblitsDir."uniprot20_current"; 
 my $pdb_full = $localHhblitsDir."pdb_full";
+my $hhblits_check_suffix = "_hhm_db";
 #cluster: "/var/tmp/rost_db/data/hhblits/uniprot20_current"; #jobtest: "/mnt/project/rost_db/data/hhblits/uniprot20_current"; # database for first HHblits run (to build the profile)
 my $hit_list = 10000; # used for -B (maximum number of alignments in alignment list) and -Z (maximum number of lines in summary hit list) parameters in the hhr output of the second HHblits run (against pdb_full)
 my $hhblits_path = "/usr/bin/hhblits";
 my $cache_path = "/usr/bin/ppc_store";
+my $get_path = "/usr/bin/ppc_fetch";
+my $pp_hhblits_hhr = "hhblits_hhr";
+my $pp_hhblits_hhm = "hhblits_hhm";
+my $pp_hhblits_a3m = "hhblits_a3m";
 my $parser_path = $rootDir."src/pssh2/parse_hhr.pl";
 my $md5script_path = $rootDir."src/pssh2/fasta_to_md5.rb";
 
 # Parse command line parameter
 my($m, $i, $t, $o, $h, $d);
-$o = getcwd;
-$t = getcwd;
+$o = getcwd();
+$t = getcwd();
 
 my $args_ok = GetOptions(
     'i=s' => \$i, #file name of input sequence
@@ -46,28 +52,32 @@ if(!($m || $i)){
     exit;	
 }
 if (defined $i && defined $m){
-    print STDERR "WARNING: You gave -i $i and -m $m, will only use $i \n";
-    undef $m;
+    print STDERR "WARNING: You gave -i $i and -m $m, will assume that $m is the md5 hash of $i \n";
 }
 if (defined $m){
-    if (defined $d){
-	$queries_dir = $d;
+    unless (defined $i){
+	if (defined $d){
+	    $queries_dir = $d;
+	}
+	else {
+	    print STDERR "WARNING: You did not provide a directory for the md5sum sequence files. Will use default: $queries_dir \n";
+	}
+	$i = "$queries_dir/$m"; #input FASTA file named according to the sequence-md5sum 
+	unless (-r $i){die "Input sequence file $i not readable! \n"};
     }
-    else {
-	print STDERR "WARNING: You did not provide a directory for the md5sum sequence files. Will use default: $queries_dir \n";
-    }
-    $i = "$queries_dir/$m"; #input FASTA file named according to the sequence-md5sum 
-    unless (-r $i){die "Input sequence file $i not readable! \n"};
 }
 if (defined $i){
-    my $tmp_md5sum = "tmp_md5sum";
-    system("cat $i | $md5script_path > $tmp_md5sum");
-
-    open FILE, $tmp_md5sum or die "Generating md5sum, but couldn't open file $tmp_md5sum\n";
-    $m = join("", <FILE>);
-    close FILE;
-    unlink $tmp_md5sum;
-    chomp $m;
+    unless (defined $m){
+	my $random_string = int rand(1000);
+	my $tmp_md5sum = "md5sum.".$random_string.".tmp";
+	system("cat $i | $md5script_path > $tmp_md5sum");
+	
+	open FILE, $tmp_md5sum or die "Generating md5sum, but couldn't open file $tmp_md5sum\n";
+	$m = join("", <FILE>);
+	close FILE;
+	unlink $tmp_md5sum;
+	chomp $m;
+    }
 }
 unless (defined $t && -d $t){
     die "temporary hhblits output path $t is not a directory! \n";
@@ -76,10 +86,24 @@ unless (defined $o && -d $o){
     die "output path for parsed files $o is not a directory! \n";
 }
 
+my $pdb_age = -M $pdb_full.$hhblits_check_suffix;
+my $uniprot_age = -M $uniprot20.$hhblits_check_suffix;
+
 print ("\nRunning $m: $i \n");
-my ($cmd_hhblits1, $cmd_hhblits2, $cmd_parse_hhr, $cmd_ppc, $cmd_ppc2) = init($i, $m, $t, $o);
-my $exit_run_hhblits1 = run_hhblits1($cmd_hhblits1, $cmd_ppc);
-my $exit_run_hhblits2 = run_hhblits2($exit_run_hhblits1, $cmd_hhblits2, $cmd_ppc2);
+my ($cmd_hhblits1, $cmd_hhblits2, $cmd_parse_hhr, $cmd_ppc, $cmd_ppg, $cmd_ppc2, $cmd_ppg2) = init($i, $m, $t, $o);
+my $exit_pp_get = run_get_ppcache($t, $uniprot_age, $pdb_age, $cmd_ppg, $cmd_ppg2);
+my ($exit_run_hhblits1, $exit_run_hhblits2);
+if ($exit_pp_get == 0){
+    $exit_run_hhblits2 = 0;
+}
+elsif ($exit_pp_get == 1){
+    $exit_run_hhblits1 = 0;
+    $exit_run_hhblits2 = run_hhblits2($exit_run_hhblits1, $cmd_hhblits2, $cmd_ppc2);
+}
+elsif ($exit_pp_get == 2){
+    $exit_run_hhblits1 = run_hhblits1($cmd_hhblits1, $cmd_ppc);
+    $exit_run_hhblits2 = run_hhblits2($exit_run_hhblits1, $cmd_hhblits2, $cmd_ppc2);
+}
 my $exit_parse_hhr = parse_hhr($exit_run_hhblits2, $cmd_parse_hhr);
 
 print "\nFinished successfully $m.\n";
@@ -104,7 +128,7 @@ OR
 =head 2 Subroutine init
 Initiates parameters and HHblits system calls
 input: ($in, $out)
-output: ($cmd_hhblits1, $cmd_hhblits2, $cmd_parse_hhr)
+output: ($cmd_hhblits1, $cmd_hhblits2, $cmd_parse_hhr, $cmd_ppc, $cmd_ppg, $cmd_ppc2, $cmd_ppg2)
 =cut
 sub init {
     my ($i, $m, $t, $o) = @_;
@@ -114,7 +138,6 @@ sub init {
     my $ohhr1 = $t."/".$m."-uniprot20.hhr";
     my $oa3m1 = $t."/".$m."-uniprot20.a3m";
     my $ohhr = $t."/".$m."-uniprot20-pdb_full.hhr";
-    my $oa3m = $t."/".$m."-uniprot20-pdb_full.a3m";
     my $parsed_ohhrs = $o."/".$m.".pssh2";
 #    my $cmd_hhblits1 = "(/usr/bin/time ".$hhblits_path." -i $i -d $uniprot20 -ohhm $ohhm -oa3m $oa3m1 -o $ohhr1) 2>> $time_log"."_hhblits1";
 #    my $cmd_hhblits2 = "(/usr/bin/time ".$hhblits_path." -i $ohhm -d $pdb_full -n 1 -B $hit_list -Z $hit_list -o $ohhr) 2>> $time_log"."_hhblits2"; 
@@ -122,10 +145,50 @@ sub init {
     my $cmd_hhblits1 = $hhblits_path." -cpu 1 -i $i -d $uniprot20 -ohhm $ohhm -oa3m $oa3m1 -o $ohhr1";
     my $cmd_hhblits2 = $hhblits_path." -cpu 1 -i $ohhm -d $pdb_full -n 1 -B $hit_list -Z $hit_list -o $ohhr"; 
     my $cmd_parse_hhr = $parser_path." -i $ohhr -m $m -o $parsed_ohhrs";
-    my $cmd_ppc = $cache_path." --seqfile $i --method=hhblits,db=uniprot20,res_hhblits_hhm=$ohhm,res_hhblits_hhr=$ohhr1,res_hhblits_a3m=$oa3m1";
-    my $cmd_ppc2 = $cache_path." --seqfile $i --method=hhblits,db=pdb_full,res_hhblits_hhr=$ohhr";
+    my $cmd_ppc = $cache_path." --seqfile $i --method=hhblits,db=uniprot20,res_$pp_hhblits_hhm=$ohhm,res_$pp_hhblits_hhr=$ohhr1,res_hhblits_a3m=$oa3m1";
+    my $cmd_ppg = $get_path." --seqfile $i --method=hhblits,db=uniprot20 -o $t -p $m";
+    my $cmd_ppc2 = $cache_path." --seqfile $i --method=hhblits,db=pdb_full,res_$pp_hhblits_hhr=$ohhr";
+    my $cmd_ppg2 = $get_path." --seqfile $i --method=hhblits,db=pdb_full -o $t -p $m";
     
-    return ($cmd_hhblits1, $cmd_hhblits2, $cmd_parse_hhr, $cmd_ppc, $cmd_ppc2); 
+    return ($cmd_hhblits1, $cmd_hhblits2, $cmd_parse_hhr, $cmd_ppc, $cmd_ppg, $cmd_ppc2, $cmd_ppg2); 
+
+}
+
+#-------------------------------------------------------------------------------
+=head 2 Subroutine 
+check the PredictProtein cache and gets results from there
+input: ($t, $uniprot_time, $pdb_time, $cmd_ppg, $cmd_ppg2)
+output: cache get success status (0: got the pdb_full run, 1: got the uniprot20 run, 2: got nothing)
+=cut
+sub run_get_ppcache {
+    
+    my ($t, $uniprot_age, $pdb_age, $cmd_ppg, $cmd_ppg2) = @_;
+    my $exitVal;
+    print "\nChecking the PP chache...\n";   
+
+    # first check if we get the pdb result from the cache
+    system($cmd_ppg2);
+    my $pp_ohhr = $t."/".$m.".".$pp_hhblits_hhr;
+    if (-e $pp_ohhr && -M $pp_ohhr < $pdb_age){
+	my $ohhr = $t."/".$m."-uniprot20-pdb_full.hhr";
+	rename $pp_ohhr, $ohhr;
+	$exitVal = 0;
+    }
+    else {
+	my $pp_ohhm = $t."/".$m.".".$pp_hhblits_hhm;
+	# now check if we get the uniprot20 result from the cache
+	system($cmd_ppg);
+	if (-e $pp_ohhm && -M $pp_ohhm < $uniprot_age){
+	    my $ohhm = $t."/".$m."-uniprot20.hhm";
+	    rename $pp_ohhm, $ohhm;
+	    $exitVal = 1;
+	}
+	else {
+	    $exitVal = 2;
+	}
+    }
+
+    return $exitVal;
 
 }
